Move partition dtypes method to base class.
